+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add dialog when exporting instances for Modrinth showing the override files [#785]
 - Support installing servers from CurseForge packs
 - Support searching CurseForge for NeoForge mods
+- Convert packs from one ATLauncher pack to another, used for migration in ATLauncher backend
 
 ### Fixes
 - Exporting an instance not working if the directory doesn't exist